# development/src/main/java/gdcau/backend/development/models/University.java
package gdcau.backend.development.models;

import jakarta.persistence.Id; // или javax.persistence.Id

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
public class University {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    
    private String name;
    private String logoUrl;
    private String websiteUrl;
    private String country;
}